refactor: extract subscribable module
