+++ packages/tinacms/src/tina-cms.ts
@@ -36,6 +36,7 @@ import { Form } from '@tinacms/forms'
 import { MediaManager, MediaStore, MediaUploadOptions } from './media'
 import { Theme } from '@tinacms/styles'
 import { Alerts } from './tina-cms/alerts'
+import { subscribable, Sub } from './subscribable'
 
 export declare type SidebarPosition = 'fixed' | 'float' | 'displace' | 'overlay'
 
@@ -120,77 +121,4 @@ class DummyMediaStore implements MediaStore {
       filename: file.name,
     }))
   }
-}
-
-function subscribable<T extends object>(target: T): T & Sub<T> {
-  const subscribableTarget = target as T & Sub<T>
-
-  const subscribers: {
-    cb: SubscriptionCallback<T>
-    subscription: Subscription<T>
-  }[] = []
-
-  function subscribe(
-    cb: SubscriptionCallback<T>,
-    subscription: Subscription<T>
-  ) {
-    subscribers.push({
-      cb,
-      subscription,
-    })
-
-    return () => unsubscribe(cb)
-  }
-
-  function unsubscribe(removeCB: SubscriptionCallback<T>) {
-    const index = subscribers.findIndex(({ cb }) => cb === removeCB)
-    subscribers.splice(index, 1)
-  }
-
-  function notifyChangeFor(prop: keyof (T & Subscription<T>), obj: T) {
-    subscribers.forEach(({ subscription, cb }) => {
-      if (!subscription || subscription[prop]) {
-        cb(obj)
-      }
-    })
-  }
-
-  const p = new Proxy<T & Sub<T>>(subscribableTarget, {
-    get: function(obj, prop: keyof (T & Subscription<T>)) {
-      if (prop === 'subscribe') {
-        return subscribe
-      }
-      if (prop === 'unsubscribe') {
-        return unsubscribe
-      }
-
-      return obj[prop]
-    },
-    set: function(obj, prop: keyof (T & Subscription<T>), value) {
-      if (prop === 'subscribe' || prop === 'unsubscribe') {
-        return true
-      }
-
-      obj[prop] = value
-      notifyChangeFor(prop, { ...obj })
-
-      return true
-    },
-  })
-
-  return p
-}
-
-export interface Sub<T> {
-  subscribe(
-    cb: SubscriptionCallback<T>,
-    subscription?: Subscription<T>
-  ): Unsubscribe
-  unsubscribe(cb: SubscriptionCallback<T>): void
-}
-
-export type SubscriptionCallback<T> = (t: T) => void
-export type Unsubscribe = () => void
-export type Subscription<T> = {
-  [P in keyof T]?: boolean
 }
